Slightly nicer OpenCL scaffolding.
